+++ lib/logger_console.dart
@@ -1,11 +1,36 @@
+/// Created by nghinv on Tue Jun 07 2022
+
 library logger_console;
 
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
+
+part 'bloc_event.dart';
+
+enum LogType {
+  info,
+  group,
+  groupCollapsed,
+  groupEnd,
+}
+
+class LogConfig {
+  final LogType? type;
+  final String? color;
+  final String? fontWeight;
+  final int? fontSize;
+
+  LogConfig({
+    this.type = LogType.info,
+    this.color,
+    this.fontWeight,
+    this.fontSize,
+  });
+}
 
 class Console {
   static WebSocketChannel? _channel;
@@ -20,7 +45,7 @@ class Console {
   ///
   /// default = true when run in debug mode
   /// default = false when run in release mode
-  static bool logEnable = kDebugMode ? true : false;
+  static bool enableLog = kDebugMode ? true : false;
 
   /// when uri != null, use uri to connect socket
   static String? uri;
@@ -44,6 +69,7 @@ class Console {
     return Uri.parse(uriString);
   }
 
+  /// Get device info
   static Future getDeviceInfo() async {
     final deviceInfoPlugin = DeviceInfoPlugin();
     final deviceInfo = await deviceInfoPlugin.deviceInfo;
@@ -120,7 +146,6 @@ class Console {
       'type': 'fromApp',
       'clientInfo': clientInfo,
     };
-
     _channel!.sink.add(json.encode(dataSending));
 
     if (data != null) {
@@ -144,8 +169,8 @@ class Console {
   ///
   /// Console.log(json.decode(response));
   /// ```
-  static log(dynamic param, [dynamic params]) {
-    if (logEnable == false) {
+  static log(dynamic param, [dynamic params, LogConfig? logConfig]) {
+    if (enableLog == false) {
       return;
     }
 
@@ -164,6 +189,12 @@ class Console {
 
     final Map<String, dynamic> dataSending = {
       'type': 'fromApp',
+      'config': {
+        'type': logConfig?.type?.name ?? LogType.info.name,
+        'color': logConfig?.color,
+        'fontWeight': logConfig?.fontWeight,
+        'fontSize': logConfig?.fontSize,
+      },
       'data': {'param': jsonParam, 'params': jsonParams}
     };
 
@@ -179,5 +210,95 @@ class Console {
     }
 
     _channel!.sink.add(json.encode(dataSending));
+  }
+
+  static group(dynamic message, [LogConfig? params]) {
+    log(message, null, params ?? LogConfig(type: LogType.group));
+  }
+
+  static groupCollapsed(dynamic message, [LogConfig? params]) {
+    log(message, null, params ?? LogConfig(type: LogType.groupCollapsed));
+  }
+
+  static groupEnd() {
+    log(null, null, LogConfig(type: LogType.groupEnd));
+  }
+
+  /// Bloc to log
+  ///
+  /// [Bloc event]
+  /// ```dart
+  /// abstract class ExampleEvent extends BlocBaseEvent {}
+  ///
+  /// class ExampleEventAdd extends ExampleEvent {
+  ///   final ExampleState exampleState;
+  ///
+  ///   ExampleEventAdd(this.exampleState);
+  ///
+  ///   @override
+  ///   eventToPayload() {
+  ///     return exampleState.toJson();
+  ///   }
+  ///
+  ///   @override
+  ///   List<Object?> get props => [exampleState];
+  /// }
+  /// ```
+  ///
+  /// [BlocObserver]
+  /// ```dart
+  /// import 'package:flutter_bloc/flutter_bloc.dart';
+  /// import 'package:logger/logger.dart';
+  ///
+  /// class AppBlocObserver extends BlocObserver {
+  ///   @override
+  ///   void onTransition(Bloc bloc, Transition transition) async {
+  ///     super.onTransition(bloc, transition);
+  ///     String eventType = transition.event.runtimeType.toString();
+  ///     var currentState;
+  ///     var nextState;
+  ///     var eventPayload;
+  ///
+  ///     try {
+  ///       currentState = transition.currentState.map((e) => e.toJson()).toList();
+  ///     } catch (e) {
+  ///       currentState = transition.currentState.toString();
+  ///     }
+  ///     try {
+  ///       nextState = transition.nextState.map((e) => e.toJson()).toList();
+  ///     } catch (e) {
+  ///       nextState = transition.nextState.toString();
+  ///     }
+  ///     try {
+  ///       eventPayload = transition.event.eventToPayload();
+  ///     } catch (e) {
+  ///       eventPayload = transition.event.toString();
+  ///     }
+  ///
+  ///     Console.logBloc(
+  ///       currentState,
+  ///       nextState,
+  ///       BlocEvent(type: eventType, payload: eventPayload),
+  ///     );
+  ///   }
+  ///
+  ///   @override
+  ///   void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  ///     super.onError(bloc, error, stackTrace);
+  ///     Console.log('onError', error);
+  ///   }
+  /// }
+  ///```
+  static logBloc(dynamic currentState, dynamic nextState, BlocEvent event) {
+    DateTime time = DateTime.now();
+    log({
+      'type': 'bloc',
+      'data': {
+        'preState': currentState,
+        'nextState': nextState,
+        'event': event.toJson(),
+        'time': time.toString(),
+      },
+    });
   }
 }